Feat: Implementacion DTO , login y registro y algunas correcciones

+++ src/main/java/com/example/ProyectoSpringAndresCastellanos/Model/MovimientoDetalle.java
@@ -21,5 +21,5 @@ public class MovimientoDetalle {
     @JoinColumn(name = "producto_id", nullable = false)
     private Producto producto;
     @Column(nullable = false)
-    private Long cantidad;
+    private int cantidad;
 }
